+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Performance events API via `OnPerfEvent` callback.
+
 ## 3.10.0 (2024-05-13)
 
 - changed: Replace all instances of TouchableOpacity and TouchableWithoutFeedback with their improved Edge implementations